cleanup common types and put them in a seperate folder
